Added linux specific installs

+++ setup.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
+if [ -f .bashrc ]; then
 rm -r ~/.bashrc \
     ~/.bash_profile \
     ~/.aliases \
     ~/.functions \
     ~/.bash_prompt
+fi
 
 ln -s ~/dotfiles/.bashrc ~/
 ln -s ~/dotfiles/.bash_profile ~/
